Clean up modal logic
